refactor in response to discussions
